+++ Assets/Scripts/TableRouteMap.cs
@@ -4,28 +4,28 @@ using UnityEngine;
 
 public class TableRouteMap
 {
-    public static IDictionary<string, int[,]> TableRouteDictionary = new Dictionary<string, int[,]>()
+    public static IDictionary<string, float[,]> TableRouteDictionary = new Dictionary<string, float[,]>()
     {
-        { "Table1Route1", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table1Route2", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table1Route3", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table1Route4", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table2Route1", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table2Route2", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table2Route3", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table2Route4", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table3Route1", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table3Route2", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table3Route3", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table3Route4", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table4Route1", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table4Route2", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table4Route3", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table4Route4", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table5Route1", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table5Route2", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table5Route3", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
-        { "Table5Route4", new[,] { { 1, 2 }, { 3, 4 }, { 5, 6 }, { 7, 8 } } },
+        { "Table1Route1", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table1Route2", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table1Route3", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table1Route4", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table2Route1", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table2Route2", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table2Route3", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table2Route4", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table3Route1", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table3Route2", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table3Route3", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table3Route4", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table4Route1", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table4Route2", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table4Route3", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table4Route4", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table5Route1", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table5Route2", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table5Route3", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
+        { "Table5Route4", new[,] { { 1f, 2f }, { 3f, 4f }, { 5f, 6f }, { 7f, 8f } } },
     };
 
 }
